requirements.txt
annotated-types==0.7.0
anyio==4.12.1
asgiref==3.11.0
boto3==1.42.64
botocore==1.42.64
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.5
distro==1.9.0
Django==5.2.8
django-cors-headers==4.9.0
django-environ==0.12.0
django-gmailapi-backend==0.3.3
django-ses==4.7.2
djangorestframework==3.16.1
google-api-core==2.30.0
google-api-python-client==2.192.0
google-auth==2.49.0
google-auth-httplib2==0.3.0
googleapis-common-protos==1.73.0
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
idna==3.11
jiter==0.13.0
jmespath==1.1.0
lxml==6.0.2
openai==2.17.0
pillow==12.0.0
proto-plus==1.27.1
protobuf==6.33.5
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pyparsing==3.3.2
pypdf==6.6.2
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-docx==1.2.0
requests==2.32.5
rsa==4.9.1
s3transfer==0.16.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.4
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.5.0
psycopg[binary]==3.3.3
pytest==9.0.2
pytest-django==4.12.0
pytest-sugar==1.1.1
